fix:deletion of users and added comments

+++ src/company-products/company-products.service.ts
@@ -4,19 +4,18 @@ import { UpdateCompanyProductDto } from './dto/update-company-product.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CompanyProduct } from './entities/company-product.entity';
 import { Repository } from 'typeorm';
-import { NODATA } from 'dns';
 
 @Injectable()
 export class CompanyProductsService {
 
   constructor(@InjectRepository(CompanyProduct) private companyProductRepo: Repository<CompanyProduct>) { }
 
-  async create(createCompanyProductDto: CreateCompanyProductDto) {
+  async create(createCompanyProductDto: CreateCompanyProductDto): Promise<CompanyProduct> {
     try {
-      // console.log('request ', req);
+      // Create company product
       const companyProduct = new CompanyProduct(createCompanyProductDto)
-
       return this.companyProductRepo.save(companyProduct);
+
     } catch (error) {
       throw error
     }
@@ -24,6 +23,7 @@ export class CompanyProductsService {
 
   async findAll(): Promise<CompanyProduct[]> {
     try {
+      // Get all products
       const products = await this.companyProductRepo.find()
       return products
     } catch (error) {
@@ -32,42 +32,53 @@ export class CompanyProductsService {
   }
 
   async findOne(id: string): Promise<CompanyProduct> {
+    // Check if the product exists
     const product = await this.companyProductRepo.findOneBy({ id })
     if (!product) throw new NotFoundException("Company product not found");
 
     return product;
   }
 
-  async update( id: string, updateCompanyProductDto: UpdateCompanyProductDto) {
+  async update(id: string, updateCompanyProductDto: UpdateCompanyProductDto): Promise<CompanyProduct> {
     try {
-      // console.log('Req ', req);
+      // Directly update the company product
       const isUpdated = await this.companyProductRepo.update(id, updateCompanyProductDto)
       if (isUpdated.affected === 0) throw new NotFoundException("Company product not found")
-        
-        const updated = await this.companyProductRepo.findOneBy({ id })
-        if (!updated) throw new NotFoundException("Company product not found")
-          return updated;
-        
-      } catch (error) {
-        throw error 
-      }
-  }
 
-  async remove(id: string):Promise<Boolean> {
-    try {
-      const isDeleted = await this.companyProductRepo.delete({id})
-      return isDeleted.affected !==0
+      // Check if the updated product is there
+      const updated = await this.companyProductRepo.findOneBy({ id })
+      if (!updated) throw new NotFoundException("Company product not found");
+      return updated;
+
     } catch (error) {
-      throw error 
+      throw error
     }
   }
 
-  async removeAll():Promise<Boolean> {
+  async remove(id: string): Promise<Boolean> {
     try {
-      const areDeleted = await this.companyProductRepo.delete({})
-      return areDeleted.affected !==0
+      // Check if user exists
+      const user = await this.companyProductRepo.findOneBy({ id })
+      if (!user) throw new NotFoundException("Company product not found")
+
+      // Delete the product
+      const isDeleted = await this.companyProductRepo.delete({ id })
+      // Check true or false for deletion
+
+      return isDeleted.affected !== 0
     } catch (error) {
-      throw error 
+      throw error
+    }
+  }
+
+  async removeAll(): Promise<Boolean> {
+    try {
+      // deleted all products
+      const areDeleted = await this.companyProductRepo.delete({})
+      // Check true or false for deletion
+      return areDeleted.affected !== 0
+    } catch (error) {
+      throw error
     }
   }
 }
