+++ src/saving-product/saving-product.service.ts
@@ -12,8 +12,8 @@ export class SavingProductService {
   ) { }
   async create(createSavingProductDto: CreateSavingProductDto): Promise<SavingProduct> {
     try {
+      // Create saving product
       const savingProduct = new SavingProduct(createSavingProductDto)
-
       return await this.savingProductRepo.save(savingProduct)
 
     } catch (error) {
@@ -24,6 +24,7 @@ export class SavingProductService {
 
   async findAll(): Promise<SavingProduct[]> {
     try {
+      // Get all saving products
       const savingProducts = await this.savingProductRepo.find()
       return savingProducts;
     } catch (error) {
@@ -31,8 +32,9 @@ export class SavingProductService {
     }
   }
 
-  async findOne(id: string):Promise<SavingProduct> {
+  async findOne(id: string): Promise<SavingProduct> {
     try {
+      // Check if savingProduct exists
       const savingProduct = await this.savingProductRepo.findOneBy({ id })
       if (!savingProduct) throw new NotFoundException("Saving product not found")
       return savingProduct;
@@ -42,15 +44,19 @@ export class SavingProductService {
     }
   }
 
-  async update(id: string, updateSavingProductDto: UpdateSavingProductDto):Promise<SavingProduct> {
+  async update(id: string, updateSavingProductDto: UpdateSavingProductDto): Promise<SavingProduct> {
     try {
+      // Directly update the savingProduct
       const toBeUpdated = await this.savingProductRepo.update(id, updateSavingProductDto)
+
+      // Check if saving Product was updated 
       if (toBeUpdated.affected === 0) {
         throw new NotFoundException("Saving product not found")
       }
 
+      // Return the updated savingProduct
       const updated = await this.savingProductRepo.findOneBy({ id });
-      if(!updated) throw new NotFoundException("Saving product not found");
+      if (!updated) throw new NotFoundException("Saving product not found");
 
       return updated;
     } catch (error) {
@@ -58,10 +64,14 @@ export class SavingProductService {
     }
   }
 
-  async remove(id: string):Promise<Boolean> {
+  async remove(id: string): Promise<Boolean> {
     try {
-      const result = await this.savingProductRepo.delete({id})
-      
+      // Check if the savingProduct exists
+      const product = await this.savingProductRepo.findOneBy({ id })
+      if (!product) throw new NotFoundException("Saving product not found")
+
+      // Delete the product
+      const result = await this.savingProductRepo.delete({ id })
       return result.affected !== 0;
 
     } catch (error) {
@@ -69,10 +79,10 @@ export class SavingProductService {
     }
   }
 
-  async removeAll():Promise<Boolean> {
+  async removeAll(): Promise<Boolean> {
     try {
+      // Deleted all saving products
       const result = await this.savingProductRepo.delete({})
-      
       return result.affected !== 0;
     } catch (error) {
       throw error
